하드웨어 브링업 중복 관련 수정
