--- src/location_manager/launch/location_manager_with_hardware.launch.py
@@ -1,96 +1,0 @@
-#!/usr/bin/env python3
-
-import os
-from ament_index_python.packages import get_package_share_directory
-from launch import LaunchDescription
-from launch.actions import DeclareLaunchArgument
-from launch.substitutions import LaunchConfiguration
-from launch_ros.actions import Node
-from launch.actions import IncludeLaunchDescription
-from launch.launch_description_sources import PythonLaunchDescriptionSource
-
-def generate_launch_description():
-    # 환경 변수 설정
-    TURTLEBOT3_MODEL = os.environ.get('TURTLEBOT3_MODEL', 'waffle_pi_4wheel')
-    
-    # Launch 파라미터
-    use_sim_time = LaunchConfiguration('use_sim_time', default='false')
-    lidar_port = LaunchConfiguration('lidar_port', default='/dev/ttyUSB0')
-    usb_port = LaunchConfiguration('usb_port', default='/dev/ttyACM0')
-    
-    return LaunchDescription([
-        # Launch 파라미터 선언
-        DeclareLaunchArgument(
-            'use_sim_time',
-            default_value='false',
-            description='Use simulation (Gazebo) clock if true'),
-        
-        DeclareLaunchArgument(
-            'lidar_port',
-            default_value='/dev/ttyUSB0',
-            description='Connected USB port with LIDAR sensor'),
-        
-        DeclareLaunchArgument(
-            'usb_port',
-            default_value='/dev/ttyACM0',
-            description='Connected USB port with OpenCR'),
-        
-        # 1. 하드웨어 브링업 (모터 + 센서 + TF)
-        IncludeLaunchDescription(
-            PythonLaunchDescriptionSource([os.path.join(
-                get_package_share_directory('turtlebot3_hardware_bringup'), 'launch', 'hardware.launch.py')]),
-            launch_arguments={
-                'lidar_port': lidar_port,
-                'usb_port': usb_port,
-                'use_sim_time': use_sim_time
-            }.items(),
-        ),
-        
-        # 2. Cartographer (SLAM) - 직접 노드 실행 (하드웨어 제외)
-        Node(
-            package='cartographer_ros',
-            executable='cartographer_node',
-            name='cartographer_node',
-            output='screen',
-            parameters=[{
-                'use_sim_time': use_sim_time,
-                'configuration_basename': 'turtlebot3_' + TURTLEBOT3_MODEL + '.lua'
-            }],
-            arguments=[
-                '-configuration_directory', os.path.join(
-                    get_package_share_directory('turtlebot3_cartographer'), 'config'),
-                '-configuration_basename', 'turtlebot3_' + TURTLEBOT3_MODEL + '.lua'
-            ]),
-        
-        # 3. Cartographer Occupancy Grid Node
-        Node(
-            package='cartographer_ros',
-            executable='cartographer_occupancy_grid_node',
-            name='cartographer_occupancy_grid_node',
-            output='screen',
-            parameters=[{'use_sim_time': use_sim_time}],
-            arguments=['-resolution', '0.05', '-publish_period_sec', '1.0']),
-        
-        # 4. Navigation2 - 직접 노드 실행 (하드웨어 제외)
-        IncludeLaunchDescription(
-            PythonLaunchDescriptionSource([os.path.join(
-                get_package_share_directory('nav2_bringup'), 'launch', 'bringup_launch.py')]),
-            launch_arguments={
-                'use_sim_time': use_sim_time,
-                'map': os.path.join(
-                    get_package_share_directory('turtlebot3_navigation2'),
-                    'map', 'map.yaml'),
-                'params_file': os.path.join(
-                    get_package_share_directory('turtlebot3_navigation2'),
-                    'param', 'humble', TURTLEBOT3_MODEL + '.yaml')
-            }.items(),
-        ),
-        
-        # 5. 위치 관리 노드
-        Node(
-            package='location_manager',
-            executable='location_manager',
-            name='location_manager',
-            output='screen',
-            env={'TURTLEBOT3_MODEL': TURTLEBOT3_MODEL}),
-    ])
