+++ src/turtlebot3_cartographer/launch/cartographer_only.launch.py
@@ -6,8 +6,6 @@ from launch import LaunchDescription
 from launch.actions import DeclareLaunchArgument
 from launch.substitutions import LaunchConfiguration
 from launch_ros.actions import Node
-from launch.actions import IncludeLaunchDescription
-from launch.launch_description_sources import PythonLaunchDescriptionSource
 
 def generate_launch_description():
     # 환경 변수 설정
@@ -15,8 +13,6 @@ def generate_launch_description():
     
     # Launch 파라미터
     use_sim_time = LaunchConfiguration('use_sim_time', default='false')
-    lidar_port = LaunchConfiguration('lidar_port', default='/dev/ttyUSB0')
-    usb_port = LaunchConfiguration('usb_port', default='/dev/ttyACM0')
     
     return LaunchDescription([
         # Launch 파라미터 선언
@@ -25,28 +21,7 @@ def generate_launch_description():
             default_value='false',
             description='Use simulation (Gazebo) clock if true'),
         
-        DeclareLaunchArgument(
-            'lidar_port',
-            default_value='/dev/ttyUSB0',
-            description='Connected USB port with LIDAR sensor'),
-        
-        DeclareLaunchArgument(
-            'usb_port',
-            default_value='/dev/ttyACM0',
-            description='Connected USB port with OpenCR'),
-        
-        # 1. 하드웨어 브링업 (모터 + 센서 + TF)
-        IncludeLaunchDescription(
-            PythonLaunchDescriptionSource([os.path.join(
-                get_package_share_directory('turtlebot3_hardware_bringup'), 'launch', 'hardware.launch.py')]),
-            launch_arguments={
-                'lidar_port': lidar_port,
-                'usb_port': usb_port,
-                'use_sim_time': use_sim_time
-            }.items(),
-        ),
-        
-        # 2. Cartographer (SLAM) - 하드웨어 노드 제외
+        # Cartographer (SLAM) - 하드웨어 제외
         Node(
             package='cartographer_ros',
             executable='cartographer_node',
@@ -62,7 +37,7 @@ def generate_launch_description():
                 '-configuration_basename', 'turtlebot3_' + TURTLEBOT3_MODEL + '.lua'
             ]),
         
-        # 3. Cartographer Occupancy Grid Node
+        # Cartographer Occupancy Grid Node
         Node(
             package='cartographer_ros',
             executable='cartographer_occupancy_grid_node',
